Fix a few compile errors

+++ simde/arm/neon/uint32x2.h
@@ -59,7 +59,7 @@ simde_vaba_u32(simde_uint32x2_t a, simde_uint32x2_t b, simde_uint32x2_t c) {
   r.mmx = _mm_add_pi32(_mm_sub_pi32(a.mmx, b.mmx), c.mmx);
 #else
   SIMDE_VECTORIZE
-  for (size_t i = 0 ; i < (sizeof(r.u16) / sizeof(r.u16[0])) ; i++) {
+  for (size_t i = 0 ; i < (sizeof(r.u32) / sizeof(r.u32[0])) ; i++) {
     r.u32[i] = (a.u32[i] - b.u32[i]) + c.u32[i];
   }
 #endif
